requirements.txt
# Core
openai>=1.3.5
langchain>=0.1.15
tiktoken
python-dotenv
langchain>=0.1.15
langchain-community>=0.0.33
langchain-core>=0.1.35

# Document parsing
PyPDF2
python-docx
docx2txt

# Web scraping
beautifulsoup4
requests

# Vector DB
chromadb>=0.4.14

# Text splitting & processing
nltk
unstructured

# Frontend
gradio>=4.14.0

# Misc
pydantic
dotenv
